Address SPI Communication Problem on ATMega2560
When using the Arduino Mega platform with the W5500 shield, there is contention between the LED signal and the SCK signal. On the Mega, the LED is mapped to digital pin 13. On the Uno, 13 is the SCK signal, but on the Mega these signals are routed to different headers altogether.

The shield still works in sketches because the SPI signals are also routed to the ISCP header, which is connected to the shield.

The logic in the bootloader is that if the SCK pin and LED pin definitions don't match, the LED pin will be initialized and will blink in the bootloader's main polling loop. On the Mega 2560, the SCK pin definition and LED pin definition _do not match_, meaning this logic was included.

Because of the routing of the signals, LED pin (Digital 13) and SCK are actually shorted together, so blinking the LED is causing contention on the SCK line.

For now, LED blinking is disabled on the ATMega2560. With this change, we can successfully upload binaries to the device over Ethernet.

Other miscellaneous improvements are also made here:

- Relocate the delay for boot-up until after SPI is initialized
- Reduce TFTP attempts required for detecting an error