Add a new menu item showing EEPROM save/erase option
A new menu item has been added in the boards.txt configuration for the Ariadne bootloaders. This allows you to manually select whether the EEPROM contents are saved or erased when the bootloader is flashed. The old default of erasing the EEPROM contents is preserved.

Documentation has been updated accordingly.

Fixes #39